+++ cli/CHANGELOG.md
@@ -1,7 +1,31 @@
 # Changelog
 
 All notable changes to this project will be documented in this file.
-## What's Changed in 4.184.0<!-- RELEASE NOTES START -->
+## What's Changed in 4.185.0<!-- RELEASE NOTES START -->
+
+### ⛰️  Features
+
+* support config-driven network proxy opt-out by [@pepicrft](https://github.com/pepicrft) in [#10334](https://github.com/tuist/tuist/pull/10334)
+### 🐛 Bug Fixes
+
+* skip result bundle upload when --inspect-mode local by [@fortmarek](https://github.com/fortmarek) in [#10401](https://github.com/tuist/tuist/pull/10401)
+
+
+
+**Full Changelog**: https://github.com/tuist/tuist/compare/4.184.2...4.185.0
+
+## What's Changed in 4.184.2<!-- RELEASE NOTES START -->
+
+### 🐛 Bug Fixes
+
+* Support  optional Buildable Folders by [@PaulTaykalo](https://github.com/PaulTaykalo) in [#9974](https://github.com/tuist/tuist/pull/9974)
+* error instead of silently ignoring mismatched test shard flags by [@fortmarek](https://github.com/fortmarek) in [#10392](https://github.com/tuist/tuist/pull/10392)
+
+
+
+**Full Changelog**: https://github.com/tuist/tuist/compare/4.184.1...4.184.2
+
+## What's Changed in 4.184.1<!-- RELEASE NOTES START -->
 
 ### ⛰️  Features
 
@@ -17,7 +41,7 @@ All notable changes to this project will be documented in this file.
 
 
 
-**Full Changelog**: https://github.com/tuist/tuist/compare/4.183.0...4.184.0
+**Full Changelog**: https://github.com/tuist/tuist/compare/4.183.0...4.184.1
 
 ## What's Changed in 4.183.0<!-- RELEASE NOTES START -->